refactor: view 파일들을 console과 web으로 분류
