+++ src/step2-index.js
@@ -1,5 +1,5 @@
-import createGameBox from "./view/layers/createGameBox.js";
-import createHeader from "./view/layers/header.js";
+import createGameBox from "./view/web/layers/createGameBox.js";
+import createHeader from "./view/web/layers/header.js";
 
 const app = async () => {
   document.addEventListener("DOMContentLoaded", async () => {
