+++ src/view/console/input/index.js
