+++ src/view/console/input/readRestartInput.js
@@ -1,5 +1,5 @@
-import { readLineAsync } from "../../utils/input.js";
-import { PROMPT_MESSAGE, LINE_BREAK } from "../../constants/message.js";
+import { readLineAsync } from '../../../utils/input.js';
+import { PROMPT_MESSAGE, LINE_BREAK } from '../../../constants/message.js';
 
 const readRestartInput = async () => {
   const userInput = await readLineAsync(
