+++ src/view/console/input/readWinningNumbersInput.js
@@ -1,11 +1,11 @@
-import { readLineAsync } from "../../utils/input.js";
-import { PROMPT_MESSAGE, LINE_BREAK } from "../../constants/message.js";
+import { readLineAsync } from '../../../utils/input.js';
+import { PROMPT_MESSAGE, LINE_BREAK } from '../../../constants/message.js';
 
 const readWinningNumbersInput = async () => {
   const userInput = await readLineAsync(
     `${LINE_BREAK}${PROMPT_MESSAGE.WINNING_NUMBER_INPUT}${LINE_BREAK}`
   );
-  return userInput.split(",").map(Number);
+  return userInput.split(',').map(Number);
 };
 
 export default readWinningNumbersInput;
