+++ src/view/console/output/index.js
