+++ src/view/console/output/printLottoNumbers.js
@@ -1,5 +1,5 @@
-import { PROMPT_MESSAGE } from "../../constants/message.js";
-import formatLottoNumbers from "../../lotto/formatLottoNumbers.js";
+import { PROMPT_MESSAGE } from "../../../constants/message.js";
+import formatLottoNumbers from "../../../lotto/formatLottoNumbers.js";
 
 const printLottoNumbers = (lottoNumberSets) => {
   console.log(`${lottoNumberSets.length}${PROMPT_MESSAGE.PURCHASE_QUANTITY}`);
