+++ src/view/web/layers/createGameBox.js
